requirements.txt
django>=4.2
djangorestframework
djangorestframework-simplejwt
argon2-cffi
pycryptodome
python-dotenv
qrcode
Pillow